backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.49
pydantic==1.10.9
python-dotenv==1.0.0
requests==2.31.0
pytz
streamlit

SpeechRecognition
pyaudio
dateparser
streamlit_js_eval
streamlit_javascript
google.generativeai

faster-whisper
python-multipart

# Optional/advanced (only install if you want AI local models)
# transformers==4.34.0
# torch   # large, only if using HF models locally

# Frontend (for part 2)
# streamlit==1.26.0

# Optional notifications
plyer==2.0.0
